test: unit tests for activity tab createGroups
