+++ src/components/AccountDrawer/MiniPortfolio/Activity/index.tsx
@@ -1,9 +1,6 @@
-import { t } from '@lingui/macro'
 import { useAccountDrawer } from 'components/AccountDrawer'
 import Column from 'components/Column'
 import { LoadingBubble } from 'components/Tokens/loading'
-import { getYear, isSameDay, isSameMonth, isSameWeek, isSameYear } from 'date-fns'
-import { TransactionStatus } from 'graphql/data/__generated__/types-and-hooks'
 import { PollingInterval } from 'graphql/data/util'
 import { atom, useAtom } from 'jotai'
 import { EmptyWalletModule } from 'nft/components/profile/view/EmptyWalletContent'
@@ -14,67 +11,7 @@ import { ThemedText } from 'theme'
 import { PortfolioSkeleton, PortfolioTabWrapper } from '../PortfolioRow'
 import { ActivityRow } from './ActivityRow'
 import { useAllActivities } from './hooks'
-import { Activity } from './types'
-
-interface ActivityGroup {
-  title: string
-  transactions: Array<Activity>
-}
-
-const sortActivities = (a: Activity, b: Activity) => b.timestamp - a.timestamp
-
-const createGroups = (activities?: Array<Activity>) => {
-  if (!activities) return undefined
-  const now = Date.now()
-
-  const pending: Array<Activity> = []
-  const today: Array<Activity> = []
-  const currentWeek: Array<Activity> = []
-  const last30Days: Array<Activity> = []
-  const currentYear: Array<Activity> = []
-  const yearMap: { [key: string]: Array<Activity> } = {}
-
-  // TODO(cartcrom): create different time bucket system for activities to fall in based on design wants
-  activities.forEach((activity) => {
-    if (activity.status === TransactionStatus.Pending) {
-      pending.push(activity)
-      return
-    }
-    const addedTime = activity.timestamp * 1000
-
-    if (isSameDay(now, addedTime)) {
-      today.push(activity)
-    } else if (isSameWeek(addedTime, now)) {
-      currentWeek.push(activity)
-    } else if (isSameMonth(addedTime, now)) {
-      last30Days.push(activity)
-    } else if (isSameYear(addedTime, now)) {
-      currentYear.push(activity)
-    } else {
-      const year = getYear(addedTime)
-
-      if (!yearMap[year]) {
-        yearMap[year] = [activity]
-      } else {
-        yearMap[year].push(activity)
-      }
-    }
-  })
-  const sortedYears = Object.keys(yearMap)
-    .sort((a, b) => parseInt(b) - parseInt(a))
-    .map((year) => ({ title: year, transactions: yearMap[year] }))
-
-  const transactionGroups: Array<ActivityGroup> = [
-    { title: t`Pending`, transactions: pending.sort(sortActivities) },
-    { title: t`Today`, transactions: today.sort(sortActivities) },
-    { title: t`This week`, transactions: currentWeek.sort(sortActivities) },
-    { title: t`This month`, transactions: last30Days.sort(sortActivities) },
-    { title: t`This year`, transactions: currentYear.sort(sortActivities) },
-    ...sortedYears,
-  ]
-
-  return transactionGroups.filter((transactionInformation) => transactionInformation.transactions.length > 0)
-}
+import { createGroups } from './utils'
 
 const ActivityGroupWrapper = styled(Column)`
   margin-top: 16px;
